netkvmswitch/requirements.txt
pynput
pytest
mss
numpy
opencv-python
pytest-mock
cryptography
pyOpenSSL
streamlit
pyserial